refresh

+++ includes/indicators.h
@@ -2,4 +2,4 @@
 
 #include "stdafx.h"
 
-float SMA(t_env *env, size_t period, size_t shift);
+double SMA(t_env *env, int period, int shift);
